refactor(api): ♻️ move getDishes procedure into a service
